Computer Vision/Image_Classification_App/gradio-app-1/requirements.txt
gradio
torch
torchvision
opencv-python
pillow